Generative_Annotation_Systems/requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
pydantic==2.7.1
python-multipart==0.0.9
numpy==1.26.4
deap==1.4.1
Pillow==8.0.1